钱包，充值，还欠款，工作总结，日志相关

+++ src/main/java/com/hy/salon/basic/entity/CardBalance.java
@@ -18,7 +18,7 @@ public class CardBalance extends TraceableEntity<Long> {
 
     private Double balance;
 
-    private Byte balanceTotal;
+    private Double balanceTotal;
 
     private Byte cardStatus;
